fix output
