+++ ft_output/ft_putchar_fd.c
@@ -6,11 +6,11 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/03 01:37:36 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/03 01:38:02 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/24 12:20:33 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_output.h"
 
 void	ft_putchar_fd(char c, int fd)
 {
