+++ ft_output/ft_putstr_fd.c
@@ -1,29 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_putnbr_fd.c                                     :+:      :+:    :+:   */
+/*   ft_putstr_fd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/03 01:41:16 by reasuke           #+#    #+#             */
-/*   Updated: 2023/09/24 18:28:30 by reasuke          ###   ########.fr       */
+/*   Created: 2023/09/03 01:38:08 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:21:37 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "ft_output.h"
 #include "libft.h"
 
-void	ft_putnbr_fd(int n, int fd)
+void	ft_putstr_fd(char *s, int fd)
 {
-	unsigned int	un;
-
-	if (n < 0)
-	{
-		un = -n;
-		ft_putchar_fd('-', fd);
-	}
-	else
-		un = n;
-	if (un / 10)
-		ft_putnbr_fd(un / 10, fd);
-	ft_putchar_fd('0' + un % 10, fd);
+	if (!s)
+		return ;
+	write(fd, s, ft_strlen(s));
 }
