+++ ft_printf/ft_printf.h
@@ -6,13 +6,14 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/08 14:20:27 by reasuke           #+#    #+#             */
-/*   Updated: 2024/02/24 00:05:51 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/24 12:22:11 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 
+# include "ft_output.h"
 # include "libft.h"
 # include <stdarg.h>
 # include <stdbool.h>
