+++ ft_printf/ft_vdprintf.c
@@ -6,7 +6,7 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/06 16:59:33 by reasuke           #+#    #+#             */
-/*   Updated: 2024/02/23 23:37:12 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/24 12:21:57 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
